+++ eras/shelley/test-suite/src/Test/Cardano/Ledger/Shelley/Examples/Cast.hs
@@ -61,7 +61,6 @@ import Cardano.Ledger.State (
  )
 import Cardano.Protocol.Crypto (hashVerKeyVRF)
 import Cardano.Protocol.TPraos.OCert (KESPeriod (..))
-import qualified Data.ByteString.Char8 as BS (pack)
 import qualified Data.List.NonEmpty as NE
 import Data.Maybe (fromJust)
 import qualified Data.Sequence.Strict as StrictSeq
@@ -135,7 +134,7 @@ aliceStakePoolParams =
         SJust $
           PoolMetadata
             { pmUrl = fromJust $ textToUrl 64 "alice.pool"
-            , pmHash = BS.pack "{}"
+            , pmHash = "{}"
             }
     }