+++ eras/shelley/test-suite/test/Test/Cardano/Ledger/Shelley/Serialisation/Golden/Genesis.hs
@@ -25,7 +25,6 @@ import Cardano.Protocol.Crypto (StandardCrypto, hashVerKeyVRF)
 import Cardano.Slotting.Slot (EpochSize (..))
 import Control.Monad
 import Data.Aeson hiding (Encoding)
-import qualified Data.ByteString.Char8 as BS (pack)
 import qualified Data.ListMap as LM
 import qualified Data.Map.Strict as Map
 import Data.Maybe (fromJust)
@@ -263,7 +262,7 @@ exampleShelleyGenesis =
             L.SJust $
               L.PoolMetadata
                 { L.pmUrl = fromJust $ textToUrl 64 "best.pool.com"
-                , L.pmHash = BS.pack "100ab{}100ab{}"
+                , L.pmHash = "100ab{}100ab{}"
                 }
         }
     staking =