+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Constrained/Conway/Instances/Ledger.hs
@@ -117,6 +117,7 @@ import Constrained.API.Extend hiding (Sized)
 import Constrained.API.Extend qualified as C
 import Control.DeepSeq (NFData)
 import Crypto.Hash (Blake2b_224)
+import Data.Array.Byte (ByteArray (..))
 import Data.ByteString qualified as BS
 import Data.ByteString.Short (ShortByteString)
 import Data.ByteString.Short qualified as SBS
@@ -130,6 +131,7 @@ import Data.Map.Strict qualified as Map
 import Data.Maybe
 import Data.OMap.Strict qualified as OMap
 import Data.OSet.Strict qualified as SOS
+import Data.Primitive.ByteArray (sizeofByteArray)
 import Data.Sequence (Seq)
 import Data.Sequence qualified as Seq
 import Data.Sequence.Strict (StrictSeq)
@@ -146,6 +148,7 @@ import GHC.Generics (Generic)
 import PlutusLedgerApi.V1 qualified as PV1
 import Test.Cardano.Ledger.Allegra.Arbitrary ()
 import Test.Cardano.Ledger.Alonzo.Arbitrary ()
+import Test.Cardano.Ledger.Binary.Arbitrary (genByteArray, genShortByteString)
 import Test.Cardano.Ledger.Constrained.Conway.Instances.Basic
 import Test.Cardano.Ledger.Constrained.Conway.Instances.PParams ()
 import Test.Cardano.Ledger.Conway.Arbitrary ()
@@ -771,7 +774,7 @@ instance HasSpec ShortByteString where
   combineSpec s s' = typeSpec $ s <> s'
   genFromTypeSpec (StringSpec ls) = do
     len <- genFromSpecT ls
-    SBS.pack <$> vectorOfT len (pureGen arbitrary)
+    pureGen $ genShortByteString len
   shrinkWithTypeSpec _ = shrink
   fixupWithTypeSpec _ _ = Nothing
   cardinalTypeSpec _ = TrueSpec
@@ -787,6 +790,24 @@ instance StringLike ShortByteString where
   lengthSpec = StringSpec
   getLengthSpec (StringSpec len) = len
   getLength = SBS.length
+
+instance HasSpec ByteArray where
+  type TypeSpec ByteArray = StringSpec
+  emptySpec = mempty
+  combineSpec s s' = typeSpec $ s <> s'
+  genFromTypeSpec (StringSpec ls) = do
+    len <- genFromSpecT ls
+    pureGen $ genByteArray len
+  shrinkWithTypeSpec _ = shrink
+  fixupWithTypeSpec _ _ = Nothing
+  cardinalTypeSpec _ = TrueSpec
+  conformsTo ba (StringSpec ls) = sizeofByteArray ba `conformsToSpec` ls
+  toPreds str (StringSpec len) = satisfies (strLen_ str) len
+
+instance StringLike ByteArray where
+  lengthSpec = StringSpec
+  getLengthSpec (StringSpec len) = len
+  getLength = sizeofByteArray
 
 data StringW :: [Type] -> Type -> Type where
   StrLenW :: StringLike s => StringW '[s] Int